output_projects/New-Project_20260329_152049/requirements.txt
streamlit==1.32.0
pandas==2.2.2
numpy==1.26.4
openai==1.30.3
PyMuPDF==1.24.5
python-dotenv==1.0.1
PyYAML==6.0.1
streamlit-authenticator==0.3.2
pywin32==306; sys_platform == "win32"
python-teams-webhook==0.2.0
pytest==8.2.2
pytest-cov==5.0.0